# perf: reduce memory allocate per request (5x faster)