Hello.

When pass the m.next.ServeHTTP to function, Go runtime will allocate temporary struct as follow:

````bash
type.noalg.struct { F uintptr; R "".middleware }(SB), AX
````

To reduce the memory allocate, we use `nextfn` to store it. The benchmark as the following:

````bash
name       old time/op    new time/op    delta
Negroni-8     832ns ± 0%     150ns ± 0%   ~     (p=1.000 n=1+1)

name       old alloc/op   new alloc/op   delta
Negroni-8      592B ± 0%       64B ± 0%   ~     (p=1.000 n=1+1)

name       old allocs/op  new allocs/op  delta
Negroni-8      12.0 ± 0%       1.0 ± 0%   ~     (p=1.000 n=1+1)
````

Signed-off-by: detailyang <detailyang@gmail.com>


BTW: There is a bonus tip to reduce memory (zero allocate) which reuses the `NewResponseWriter` via sync.pool